requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
cachetools==5.2.0
certifi==2018.10.15
cfgv==3.3.1
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.5.0
distlib==0.3.5
distro==1.4.0
distro-info===0.23ubuntu1
Django==2.2.18
django-filter==2.0.0
django-rest-auth==0.9.3
djangorestframework==3.11.2
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.7.1
future==0.18.2
google-api-core==2.8.2
google-auth==2.10.0
google-cloud-core==2.3.2
google-cloud-translate==3.8.0
googleapis-common-protos==1.56.4
grpcio==1.47.0
grpcio-status==1.47.0
httplib2==0.14.0
identify==2.5.2
idna==2.7
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.0.1
MarkupSafe==1.1.0
monotonic==1.5
netifaces==0.10.4
nodeenv==1.7.0
oauthlib==2.1.0
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
platformdirs==2.5.2
pre-commit==2.20.0
proto-plus==1.20.6
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.19.0
python3-openid==3.1.0
pytz==2018.7
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.20.0
requests-oauthlib==1.0.0
requests-unixsocket==0.2.0
rsa==4.9
SecretStorage==2.3.1
simplejson==3.16.0
six==1.11.0
sqlparse==0.4.2
systemd-python==234
toml==0.10.2
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.24.2
usb-creator==0.3.7
virtualenv==20.16.2
wadllib==1.3.3
xkit==0.0.0
